+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
   * gpxdo fix --clear-times, --set-first-time, --set-last-time, --add-times, --clear-locations
   * new: Track.locate_point() and Track.add_locations()
   * split class Account into Account and DirectoryAccount
+  * Directory: the file .gpxity_config may define account settings
 
 1.6.0 release 2019-01-xx
 ------------------------